+++ backend/src/main/java/com/tamnara/backend/auth/dto/KakaoLoginResponseDto.java
@@ -1,0 +1,11 @@
+package com.tamnara.backend.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class KakaoLoginResponseDto {
+    private boolean success;
+    private String message;
+}
