+++ backend/src/main/java/com/tamnara/backend/auth/dto/KakaoUrlResponseDto.java
@@ -1,0 +1,14 @@
+package com.tamnara.backend.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.util.Map;
+
+@Data
+@AllArgsConstructor
+public class KakaoUrlResponseDto {
+    private boolean success;
+    private String message;
+    private Map<String, String> data;
+}
